Added update functions
This commit includes update functions for:
- User's Address
- User's Email
- User's Phone Number
- Emergency Contacts's Address
- Emergency Contact's Email
- Emergency Contact's Phone Number. 

There is currently an issue with the mySQL queries that isn't allowing for emergency contacts to be saved locally and addresses to be updated. The methods are still there for implementation and can be used while I work out the logic. This commit also includes more Java documentation for better understanding. You don't need to know how the functions work, you just need to know how to properly implement them and what they return (If it does). 

My Current Backlog:
- Fix mySQL queries
- Add/Delete functions for emergency contacts
- Add/Drop course registration 
- Account structure in database
- Retrieve account details from database
- Further Java Documentation   
- Efficiency improvements 

If you have any questions, concerns, or think i'm missing something that is within my scope of work - contact me.
- Jay